feat(github): share webhook PR attribution across products
