+++ posthog/api/github_webhooks/__init__.py
@@ -1,0 +1,1 @@
+"""GitHub App webhook transport and shared pull request analytics."""
