+++ posthog/api/github_webhooks/signature.py
@@ -1,0 +1,24 @@
+import hmac
+
+from posthog.models.instance_setting import get_instance_setting
+
+
+def verify_github_signature(payload: bytes, signature: str | None, secret: str) -> bool:
+    """
+    Verify the GitHub webhook signature using HMAC-SHA256.
+
+    GitHub sends a signature in the X-Hub-Signature-256 header in the format:
+    sha256=<hex_digest>
+    """
+    if not signature or not signature.startswith("sha256="):
+        return False
+
+    expected_signature = "sha256=" + hmac.digest(secret.encode("utf-8"), payload, "sha256").hex()
+
+    return hmac.compare_digest(expected_signature, signature)
+
+
+def get_github_webhook_secret() -> str | None:
+    """Get the GitHub webhook secret from instance settings."""
+    secret = get_instance_setting("GITHUB_WEBHOOK_SECRET")
+    return secret if secret else None
