+++ posthog/urls.py
@@ -1,9 +1,7 @@
-from collections.abc import Callable
 from typing import Any, cast
 from urllib.parse import urlencode, urlparse
 
 from django.conf import settings
-from django.core.cache import cache
 from django.http import HttpRequest, HttpResponse, HttpResponseRedirect, HttpResponseServerError
 from django.template import loader
 from django.urls import include, path, re_path
@@ -33,6 +31,7 @@ from posthog.api import (
     user,
 )
 from posthog.api.github_callback.views import github_oauth_callback, github_setup_callback
+from posthog.api.github_webhooks.views import github_webhook
 from posthog.api.oauth.connected_apps import ConnectedAppsViewSet
 from posthog.api.oauth.hogli_metadata import HOGLI_METADATA_PATH, HogliClientMetadataView
 from posthog.api.oauth.raycast_metadata import RAYCAST_METADATA_PATH, RaycastClientMetadataView
@@ -42,7 +41,6 @@ from posthog.api.two_factor_qrcode import CacheAwareQRGeneratorView
 from posthog.api.utils import hostname_in_allowed_url_list
 from posthog.api.web_experiment import web_experiments
 from posthog.constants import PERMITTED_FORUM_DOMAINS
-from posthog.exceptions_capture import capture_exception
 from posthog.models import User
 from posthog.models.instance_setting import get_instance_setting
 from posthog.oauth2_urls import urlpatterns as oauth2_urls
@@ -120,202 +118,6 @@ except ImportError:
     pass
 else:
     extend_api_router()
-
-
-GithubWebhookHandler = Callable[[HttpRequest, str, dict[str, Any], str], HttpResponse | None]
-
-
-def _dispatch_conversations_event(
-    request: HttpRequest, event_type: str, payload: dict[str, Any], delivery_id: str
-) -> HttpResponse:
-    from products.conversations.backend.api.github_events import dispatch_github_event
-
-    return dispatch_github_event(request, event_type, payload)
-
-
-def _dispatch_pull_request_event(
-    request: HttpRequest, event_type: str, payload: dict[str, Any], delivery_id: str
-) -> HttpResponse:
-    from products.tasks.backend.facade.webhooks import handle_pull_request_event
-
-    return handle_pull_request_event(payload)
-
-
-def _dispatch_pull_request_review_event(
-    request: HttpRequest, event_type: str, payload: dict[str, Any], delivery_id: str
-) -> HttpResponse:
-    from products.tasks.backend.facade.webhooks import handle_pull_request_review_event
-
-    return handle_pull_request_review_event(payload)
-
-
-def _dispatch_installation_event(
-    request: HttpRequest, event_type: str, payload: dict[str, Any], delivery_id: str
-) -> HttpResponse:
-    from posthog.api.github_callback.installation_events import handle_installation_event
-
-    return handle_installation_event(payload)
-
-
-def _dispatch_installation_repositories_event(
-    request: HttpRequest, event_type: str, payload: dict[str, Any], delivery_id: str
-) -> HttpResponse:
-    from posthog.api.github_callback.installation_events import handle_installation_repositories_event
-
-    return handle_installation_repositories_event(payload)
-
-
-def _dispatch_loop_triggers(request: HttpRequest, event_type: str, payload: dict[str, Any], delivery_id: str) -> None:
-    from products.tasks.backend.facade.webhooks import handle_github_event_for_loops
-
-    handle_github_event_for_loops(event_type, payload, delivery_id)
-    return None
-
-
-def _dispatch_workflow_triggers(
-    request: HttpRequest, event_type: str, payload: dict[str, Any], delivery_id: str
-) -> None:
-    from products.workflows.backend.github_workflow_events import emit_github_event
-
-    emit_github_event(event_type, payload, delivery_id)
-    return None
-
-
-# event_type -> ordered list of (handler_name, handler). Order matters only in that
-# the first handler in a bucket to return a non-None HttpResponse determines the
-# response sent back to GitHub; the pre-existing single handler in each bucket keeps
-# that slot so its response is unchanged by additive handlers registered after it.
-GITHUB_WEBHOOK_HANDLERS: dict[str, list[tuple[str, GithubWebhookHandler]]] = {
-    "issues": [
-        ("conversations", _dispatch_conversations_event),
-        ("loops", _dispatch_loop_triggers),
-        ("workflows", _dispatch_workflow_triggers),
-    ],
-    "issue_comment": [
-        ("conversations", _dispatch_conversations_event),
-        ("loops", _dispatch_loop_triggers),
-        ("workflows", _dispatch_workflow_triggers),
-    ],
-    "pull_request": [
-        ("tasks_pr_backstop", _dispatch_pull_request_event),
-        ("loops", _dispatch_loop_triggers),
-        ("workflows", _dispatch_workflow_triggers),
-    ],
-    "pull_request_review": [
-        ("tasks_pr_review", _dispatch_pull_request_review_event),
-        ("workflows", _dispatch_workflow_triggers),
-    ],
-    "installation": [
-        ("installation_lifecycle", _dispatch_installation_event),
-    ],
-    "installation_repositories": [
-        ("installation_repositories", _dispatch_installation_repositories_event),
-    ],
-    "push": [
-        ("loops", _dispatch_loop_triggers),
-        ("workflows", _dispatch_workflow_triggers),
-    ],
-}
-
-GITHUB_WEBHOOK_DELIVERY_DEDUP_TTL_SECONDS = 24 * 60 * 60
-
-
-def _is_duplicate_github_webhook_delivery(handler_name: str, delivery_id: str) -> bool:
-    """Redis-backed per-handler delivery dedup, fail-open when the cache backend errors.
-
-    Keyed per handler, not just per delivery id: one GitHub delivery legitimately fans
-    out to multiple handlers (e.g. a pull_request delivery reaches both the tasks PR
-    backstop and the Loops handler), so a delivery-wide key would starve every handler
-    but the first. This sits alongside each consumer's own dedup (e.g. the conversations
-    Celery task) rather than replacing it.
-    """
-    key = _github_webhook_delivery_key(handler_name, delivery_id)
-    try:
-        return not cache.add(key, True, timeout=GITHUB_WEBHOOK_DELIVERY_DEDUP_TTL_SECONDS)
-    except Exception:
-        logger.warning(
-            "github_webhook_dedup_cache_failed", handler=handler_name, delivery_id=delivery_id, exc_info=True
-        )
-        return False
-
-
-def _github_webhook_delivery_key(handler_name: str, delivery_id: str) -> str:
-    return f"github_webhook_delivery:{handler_name}:{delivery_id}"
-
-
-def _release_github_webhook_delivery(handler_name: str, delivery_id: str) -> None:
-    """Drop the dedup mark after a handler failed, so GitHub's redelivery of the same
-    GUID gets processed instead of silently skipped (the mark is set before the handler
-    runs, so a failure would otherwise burn the delivery for 24h)."""
-    try:
-        cache.delete(_github_webhook_delivery_key(handler_name, delivery_id))
-    except Exception:
-        logger.warning(
-            "github_webhook_dedup_release_failed", handler=handler_name, delivery_id=delivery_id, exc_info=True
-        )
-
-
-@csrf_exempt
-def github_webhook(request: HttpRequest) -> HttpResponse:
-    """Unified GitHub App webhook dispatcher.
-
-    Verifies the HMAC-SHA256 signature once, parses JSON once, then routes by
-    ``X-GitHub-Event`` to every registered product handler. Each handler runs in
-    isolation: one handler raising is logged and captured but never blocks another
-    handler or the response sent back to GitHub.
-    """
-    import json
-
-    from products.tasks.backend.facade.webhooks import get_github_webhook_secret, verify_github_signature
-
-    if request.method != "POST":
-        return HttpResponse(status=405)
-
-    secret = get_github_webhook_secret()
-    if not secret:
-        return HttpResponse("Webhook not configured", status=500)
-
-    signature = request.headers.get("X-Hub-Signature-256")
-    if not verify_github_signature(request.body, signature, secret):
-        return HttpResponse("Invalid signature", status=403)
-
-    try:
-        payload = json.loads(request.body)
-    except json.JSONDecodeError:
-        return HttpResponse("Invalid JSON", status=400)
-
-    event_type = request.headers.get("X-GitHub-Event", "")
-    delivery_id = request.headers.get("X-GitHub-Delivery", "")
-    handlers = GITHUB_WEBHOOK_HANDLERS.get(event_type, [])
-
-    logger.info(
-        "github_webhook_dispatch",
-        event_type=event_type,
-        delivery_id=delivery_id,
-        handlers_matched=[name for name, _ in handlers],
-    )
-
-    response: HttpResponse | None = None
-    for name, handler in handlers:
-        if delivery_id and _is_duplicate_github_webhook_delivery(name, delivery_id):
-            logger.info("github_webhook_handler_deduped", event_type=event_type, delivery_id=delivery_id, handler=name)
-            continue
-
-        try:
-            handler_response = handler(request, event_type, payload, delivery_id)
-        except Exception as e:
-            logger.exception(
-                "github_webhook_handler_failed", event_type=event_type, delivery_id=delivery_id, handler=name
-            )
-            capture_exception(e)
-            if delivery_id:
-                _release_github_webhook_delivery(name, delivery_id)
-            continue
-
-        if response is None and handler_response is not None:
-            response = handler_response
-
-    return response if response is not None else HttpResponse(status=200)
 
 
 @requires_csrf_token
