+++ products/tasks/backend/facade/webhooks.py
@@ -1,22 +1,6 @@
-"""
-Facade re-exports for the GitHub webhook handlers.
-
-Core's unified GitHub webhook view dispatches to these: it verifies the signature on every
-POST, routes pull-request events to the tasks handler, and fans events out to loop triggers.
-"""
+"""Task-owned GitHub PR processing and loop triggers."""
 
 from products.tasks.backend.loop_github_events import handle_github_event_for_loops
-from products.tasks.backend.webhooks import (
-    get_github_webhook_secret,
-    handle_pull_request_event,
-    handle_pull_request_review_event,
-    verify_github_signature,
-)
+from products.tasks.backend.webhooks import handle_pull_request_event, handle_pull_request_review_event
 
-__all__ = [
-    "get_github_webhook_secret",
-    "handle_github_event_for_loops",
-    "handle_pull_request_event",
-    "handle_pull_request_review_event",
-    "verify_github_signature",
-]
+__all__ = ["handle_github_event_for_loops", "handle_pull_request_event", "handle_pull_request_review_event"]
