+++ products/tasks/backend/tests/test_webhooks.py
@@ -15,6 +15,9 @@ from prometheus_client import REGISTRY
 from rest_framework.test import APIClient
 from social_django.models import UserSocialAuth
 
+from posthog.api.github_webhooks.attribution import _attribution_db_aliases, _bounded_attribution_lookup
+from posthog.api.github_webhooks.integrations import _installation_team_ids
+from posthog.api.github_webhooks.pull_requests import _PR_BODY_MAX_CHARS, _account_type
 from posthog.models.integration import Integration
 from posthog.models.organization import Organization, OrganizationMembership
 from posthog.models.team.team import Team
@@ -25,15 +28,7 @@ from products.signals.backend.implementation_pr import fetch_implementation_pr_s
 from products.signals.backend.models import SignalActorKind, SignalReport, SignalReportAssignment, SignalReportTask
 from products.tasks.backend.facade.api import find_signal_implementation_run
 from products.tasks.backend.models import Task, TaskRun, TaskThreadMessage
-from products.tasks.backend.webhooks import (
-    _PR_BODY_MAX_CHARS,
-    _account_type,
-    _attribution_db_aliases,
-    _bounded_attribution_lookup,
-    _installation_team_ids,
-    _task_run_scope_team_ids,
-    find_task_run,
-)
+from products.tasks.backend.webhooks import _task_run_scope_team_ids, find_task_run
 
 
 class TestAccountType(TestCase):
@@ -111,8 +106,8 @@ class TestGitHubPRWebhook(TestCase):
             headers={"x-hub-signature-256": signature, "x-github-event": event_type},
         )
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_webhook(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -146,8 +141,8 @@ class TestGitHubPRWebhook(TestCase):
             ("unresolvable_login", "stranger", None, "user-123"),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_attributes_to_merger(
         self, _name, merged_by_login, expected_property, expected_distinct_id, mock_capture, mock_get_secret
     ):
@@ -174,9 +169,9 @@ class TestGitHubPRWebhook(TestCase):
         self.assertEqual(call_kwargs["properties"]["pr_merged_by_id"], 583231)
         self.assertEqual(call_kwargs["properties"].get("pr_merged_by_distinct_id"), expected_property)
 
-    @patch("products.tasks.backend.webhooks.resolve_org_github_login_to_users", side_effect=RuntimeError("boom"))
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.attribution.resolve_github_login_distinct_id", side_effect=RuntimeError("boom"))
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_by_resolution_failure_keeps_webhook_successful(
         self, mock_capture, mock_get_secret, _mock_resolve
     ):
@@ -200,11 +195,11 @@ class TestGitHubPRWebhook(TestCase):
         self.assertNotIn("pr_merged_by_distinct_id", call_kwargs["properties"])
 
     @patch(
-        "products.tasks.backend.webhooks.resolve_org_github_login_to_users",
+        "posthog.api.github_webhooks.attribution.resolve_github_login_distinct_id",
         side_effect=OperationalError("canceling statement due to statement timeout"),
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_attribution_timeout_keeps_webhook_successful(self, mock_capture, mock_get_secret, _mock_resolve):
         # A slow member lookup must degrade to no attribution, never cost the delivery:
         # GitHub does not retry pull_request events and the merge side effects run after.
@@ -235,11 +230,11 @@ class TestGitHubPRWebhook(TestCase):
         self.assertIs(self.task_run.output.get("pr_merged"), True)
 
     @patch(
-        "products.tasks.backend.webhooks.resolve_org_github_login_to_users",
+        "posthog.api.github_webhooks.attribution.resolve_github_login_distinct_id",
         side_effect=OperationalError("server closed the connection unexpectedly"),
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_attribution_connection_error_is_not_counted_as_timeout(
         self, mock_capture, mock_get_secret, _mock_resolve
     ):
@@ -269,11 +264,12 @@ class TestGitHubPRWebhook(TestCase):
             _sample_value("posthog_tasks_github_webhook_attribution_total", {"outcome": "timeout"}), timeouts
         )
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture", side_effect=RuntimeError("capture down"))
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch(
+        "posthog.api.github_webhooks.pull_requests.posthoganalytics.capture", side_effect=RuntimeError("capture down")
+    )
     def test_task_backed_capture_failure_increments_drop_counter(self, _mock_capture, mock_get_secret):
-        # TaskRun.capture_event swallows the failure, so without a reported outcome the drop
-        # counter would miss the task-backed path entirely — the bulk of the traffic.
+        # Capture failures must still count as dropped events for Task-owned PRs.
         mock_get_secret.return_value = self.webhook_secret
         labels = {"analytics_event": "pr_merged", "reason": "capture_exception"}
         before = _sample_value("posthog_tasks_github_webhook_pr_event_dropped_total", labels)
@@ -288,8 +284,8 @@ class TestGitHubPRWebhook(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(_sample_value("posthog_tasks_github_webhook_pr_event_dropped_total", labels), before + 1)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_delivery_without_installation_falls_back_to_unscoped_lookup(self, mock_capture, mock_get_secret):
         # Deliveries carrying no installation block keep the legacy full-table lookup, so the
         # match must not regress. The counter is how we see how much of that traffic is left.
@@ -308,8 +304,8 @@ class TestGitHubPRWebhook(TestCase):
         self.assertEqual(mock_capture.call_args[1]["properties"]["run_id"], str(self.task_run.id))
         self.assertEqual(_sample_value("posthog_tasks_github_webhook_task_run_lookup_total", labels), before + 1)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_publishes_stream_events(self, mock_capture, mock_get_secret):
         # A live installation-progress view only learns about the merge through the stream;
         # recording output.pr_merged without publishing leaves the UI stuck on "opened".
@@ -331,8 +327,8 @@ class TestGitHubPRWebhook(TestCase):
         self.assertEqual(len(progress), 1)
         self.assertEqual(progress[0]["notification"]["params"]["label"], "Pull request merged")
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_from_fork_does_not_record_pr_merged(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         run = TaskRun.objects.create(
@@ -358,8 +354,8 @@ class TestGitHubPRWebhook(TestCase):
         run.refresh_from_db()
         self.assertEqual(run.output, {})
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_for_other_pr_on_same_branch_does_not_record_pr_merged(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         run = TaskRun.objects.create(
@@ -411,8 +407,8 @@ class TestGitHubPRWebhook(TestCase):
             ("already_terminal_run", {"wizard_config": {}}, TaskRun.Status.COMPLETED, {}, 0),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_signals_wizard_workflow_completion(
         self, _name, state, status, extra_output, expected_signals, _mock_capture, mock_get_secret
     ):
@@ -435,8 +431,8 @@ class TestGitHubPRWebhook(TestCase):
         if expected_signals:
             mock_signal.assert_called_once_with(run.id, TaskRun.Status.COMPLETED, None)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_resolves_to_active_run_when_resume_shares_pr(self, _mock_capture, mock_get_secret):
         # A resumed wizard run shares its predecessor's PR; the merge must land on the
         # live run (recording pr_merged and signaling wind-down), not the dead original.
@@ -474,8 +470,8 @@ class TestGitHubPRWebhook(TestCase):
         self.assertIs(active_run.output.get("pr_merged"), True)
         self.assertNotIn("pr_merged", terminal_run.output)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_merged_signal_failure_keeps_webhook_successful(self, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         pr_url = "https://github.com/posthog/posthog/pull/778"
@@ -517,8 +513,8 @@ class TestGitHubPRWebhook(TestCase):
             ("local_run", {"wizard_config": {}}, TaskRun.Status.IN_PROGRESS, TaskRun.Environment.LOCAL, 0),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_closed_cancels_wizard_run(
         self, _name, state, status, environment, expected_cancels, _mock_capture, mock_get_secret
     ):
@@ -548,8 +544,8 @@ class TestGitHubPRWebhook(TestCase):
                 source="pr_closed",
             )
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_closed_cancel_failure_keeps_webhook_successful(self, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         pr_url = "https://github.com/posthog/posthog/pull/781"
@@ -570,8 +566,8 @@ class TestGitHubPRWebhook(TestCase):
 
         self.assertEqual(response.status_code, 200)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_closed_without_merge_webhook(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -591,8 +587,8 @@ class TestGitHubPRWebhook(TestCase):
         call_kwargs = mock_capture.call_args[1]
         self.assertEqual(call_kwargs["event"], "pr_closed")
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_opened_webhook(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -630,8 +626,8 @@ class TestGitHubPRWebhook(TestCase):
             ("over_the_cap", _PR_BODY_MAX_CHARS + 1, True),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_body_is_capped(self, _name, body_length, expected_truncated, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -663,8 +659,8 @@ class TestGitHubPRWebhook(TestCase):
             ("closed", {"merged": True}, "merged"),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_action_records_pr_state(self, action_name, pr_fields, expected_state, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         action = "opened" if action_name == "opened_as_draft" else action_name
@@ -684,8 +680,8 @@ class TestGitHubPRWebhook(TestCase):
         assert self.task_run.output is not None
         self.assertEqual(self.task_run.output.get("pr_state"), expected_state)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_state_only_action_records_state_without_analytics(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -705,8 +701,8 @@ class TestGitHubPRWebhook(TestCase):
         assert self.task_run.output is not None
         self.assertEqual(self.task_run.output.get("pr_state"), "draft")
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_state_not_recorded_for_unclaimed_pr(self, _mock_capture, mock_get_secret):
         """A same-branch webhook for a different PR must not restate this run's PR."""
         mock_get_secret.return_value = self.webhook_secret
@@ -728,8 +724,8 @@ class TestGitHubPRWebhook(TestCase):
         assert self.task_run.output is not None
         self.assertNotIn("pr_state", self.task_run.output)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_opened_backfills_pr_url_on_branch_match(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         run = TaskRun.objects.create(
@@ -759,8 +755,8 @@ class TestGitHubPRWebhook(TestCase):
         self.assertEqual(run.state["verified_pr_urls"], [pr_url])
 
     @patch("products.tasks.backend.facade.api.posthoganalytics.feature_enabled", return_value=True)
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_opened_repairs_missing_artifact_for_existing_pr_url(
         self, mock_capture, mock_get_secret, mock_feature_enabled
     ) -> None:
@@ -790,8 +786,8 @@ class TestGitHubPRWebhook(TestCase):
             TaskThreadMessage.objects.for_team(self.team.id).filter(task=self.task, payload__pr_url=pr_url).exists()
         )
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_opened_backfills_pr_url_on_wizard_head_branch_match(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         run = TaskRun.objects.create(
@@ -820,8 +816,8 @@ class TestGitHubPRWebhook(TestCase):
         assert run.output is not None
         self.assertEqual(run.output["pr_url"], pr_url)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_opened_from_fork_does_not_backfill_pr_url(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         run = TaskRun.objects.create(
@@ -847,8 +843,8 @@ class TestGitHubPRWebhook(TestCase):
         run.refresh_from_db()
         self.assertEqual(run.output, {})
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_opened_prefers_self_driving_run_over_newer_reviewhog_run(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         head_branch = "posthog-self-driving/fix-thing-abc123"
@@ -906,8 +902,8 @@ class TestGitHubPRWebhook(TestCase):
         self.assertEqual(review_run.output, {})
         self.assertNotIn("verified_pr_urls", review_run.state or {})
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_opened_does_not_overwrite_existing_pr_url(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         existing = "https://github.com/posthog/posthog/pull/900"
@@ -939,7 +935,7 @@ class TestGitHubPRWebhook(TestCase):
             [existing, "https://github.com/posthog/posthog/pull/901"],
         )
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_invalid_signature_rejected(self, mock_get_secret):
         """Test that requests with invalid signatures are rejected."""
         mock_get_secret.return_value = self.webhook_secret
@@ -956,7 +952,7 @@ class TestGitHubPRWebhook(TestCase):
 
         self.assertEqual(response.status_code, 403)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_missing_signature_rejected(self, mock_get_secret):
         """Test that requests without signatures are rejected."""
         mock_get_secret.return_value = self.webhook_secret
@@ -972,8 +968,8 @@ class TestGitHubPRWebhook(TestCase):
 
         self.assertEqual(response.status_code, 403)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_unmatched_pr_without_installation_not_captured(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -990,7 +986,7 @@ class TestGitHubPRWebhook(TestCase):
         self.assertEqual(response.status_code, 200)
         mock_capture.assert_not_called()
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_non_pr_non_issue_event_ignored(self, mock_get_secret):
         """Test that events other than pull_request/issues/issue_comment are acknowledged but ignored."""
         mock_get_secret.return_value = self.webhook_secret
@@ -1001,7 +997,7 @@ class TestGitHubPRWebhook(TestCase):
 
         self.assertEqual(response.status_code, 200)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_ignored_pr_actions(self, mock_get_secret):
         """Test that PR actions other than opened/closed are acknowledged but ignored."""
         mock_get_secret.return_value = self.webhook_secret
@@ -1020,7 +1016,7 @@ class TestGitHubPRWebhook(TestCase):
 
     def test_webhook_secret_not_configured(self):
         """Test that webhook returns 500 if secret is not configured."""
-        with patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret", return_value=None):
+        with patch("posthog.api.github_webhooks.views.get_github_webhook_secret", return_value=None):
             payload = {"action": "closed", "pull_request": {"html_url": "https://github.com/org/repo/pull/1"}}
 
             response = self.client.post(
@@ -1037,8 +1033,8 @@ class TestGitHubPRWebhook(TestCase):
         response = self.client.get("/webhooks/github/pr/")
         self.assertEqual(response.status_code, 405)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_webhook_does_not_attribute_foreign_repo_pr_to_unrelated_run(self, mock_capture, mock_get_secret):
         # Regression: a PR opened on a repo that has no matching TaskRun must
         # not fall through to a branch-only lookup that attributes the event
@@ -1125,8 +1121,8 @@ class TestGitHubPRReviewWebhook(TestCase):
             ("unresolvable_login", "stranger", None, "user-123"),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_review_submission_attributes_to_reviewer(
         self, _name, reviewer_login, expected_property, expected_distinct_id, mock_capture, mock_get_secret
     ):
@@ -1157,8 +1153,8 @@ class TestGitHubPRReviewWebhook(TestCase):
             ("non_submitted_action", {"login": "octocat", "id": 583231, "type": "User"}, "dismissed"),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_review_events_not_captured(self, _name, reviewer, action, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -1220,8 +1216,8 @@ class TestGitHubPRWebhookResolvesSignalReports(TestCase):
             ("research", True, None, "research", SignalReport.Status.READY),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_event_falls_back_to_task_links(
         self, _name, merged, actor_kind, relationship, expected_status, _mock_capture, mock_get_secret
     ):
@@ -1301,8 +1297,8 @@ class TestGitHubPRWebhookResolvesSignalReports(TestCase):
             ),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_event_transitions_linked_report(
         self,
         _name,
@@ -1327,8 +1323,8 @@ class TestGitHubPRWebhookResolvesSignalReports(TestCase):
         self.assertEqual(self.assignment.pr_state, expected_pr_state)
         self.assertIs(self.assignment.pr_merged, merged)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_merge_without_matching_assignment_is_a_noop(self, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         self.assignment.delete()
@@ -1345,8 +1341,8 @@ class TestGitHubPRWebhookResolvesSignalReports(TestCase):
             ("closed", False, SignalReport.Status.SUPPRESSED, SignalReportAssignment.PrState.CLOSED),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_event_transitions_every_report_linked_to_the_pr(
         self, _name, merged, expected_status, expected_pr_state, _mock_capture, mock_get_secret
     ):
@@ -1385,8 +1381,8 @@ class TestGitHubPRWebhookResolvesSignalReports(TestCase):
         legacy_report.refresh_from_db()
         self.assertEqual(legacy_report.status, expected_status)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_event_does_not_transition_assignment_for_another_pr(self, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         assert self.assignment.pr_url is not None
@@ -1404,8 +1400,8 @@ class TestGitHubPRWebhookResolvesSignalReports(TestCase):
         self.assertEqual(self.assignment.pr_state, SignalReportAssignment.PrState.OPEN)
         self.assertFalse(self.assignment.pr_merged)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_event_only_updates_teams_connected_to_the_installation(self, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         other_organization = Organization.objects.create(name="Other Org")
@@ -1449,8 +1445,8 @@ class TestGitHubPRWebhookResolvesSignalReports(TestCase):
             ("www_host_trailing_slash", "https://www.github.com/posthog/posthog/pull/42/"),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pr_url_variants_match_by_repository_and_number(self, _name, stored_pr_url, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         self.assignment.pr_url = stored_pr_url
@@ -1528,8 +1524,8 @@ class TestExternalPRWebhook(TestCase):
             ("merged", "closed", True, "pr_merged"),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_external_pr_event_attributes_to_installation_team(
         self, _name, action, merged, expected_event, mock_capture, mock_get_secret
     ):
@@ -1563,8 +1559,8 @@ class TestExternalPRWebhook(TestCase):
         for key in ("pr_title", "pr_body", "pr_labels", "pr_requested_reviewers", "pr_is_draft"):
             self.assertIsNone(props[key])
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_external_pr_event_is_deduplicated_per_action(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -1577,8 +1573,8 @@ class TestExternalPRWebhook(TestCase):
         second_uuid = mock_capture.call_args_list[1][1]["uuid"]
         self.assertEqual(first_uuid, second_uuid)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_external_pr_without_resolvable_installation_is_dropped(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
 
@@ -1590,8 +1586,8 @@ class TestExternalPRWebhook(TestCase):
         self.assertEqual(response.status_code, 200)
         mock_capture.assert_not_called()
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_external_pr_unresolved_installation_increments_drop_counter(self, mock_capture, mock_get_secret):
         # The silent drop is now a counter, so a webhook-side event loss shows up as an
         # error rate instead of only a dip in the downstream capture ratio.
@@ -1606,8 +1602,10 @@ class TestExternalPRWebhook(TestCase):
         mock_capture.assert_not_called()
         self.assertEqual(_sample_value("posthog_tasks_github_webhook_pr_event_dropped_total", labels), before + 1)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture", side_effect=RuntimeError("capture down"))
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch(
+        "posthog.api.github_webhooks.pull_requests.posthoganalytics.capture", side_effect=RuntimeError("capture down")
+    )
     def test_external_pr_capture_exception_increments_drop_counter(self, _mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         labels = {"analytics_event": "pr_created", "reason": "capture_exception"}
@@ -1623,7 +1621,10 @@ class TestExternalPRWebhook(TestCase):
         # Task.github_user_integration is unindexed. Such a task lives in a team of the
         # installing user's org, so widen the scope there rather than give up and scan.
         payload = self._external_payload("opened", merged=False)
-        self.assertEqual(_task_run_scope_team_ids(payload), [self.team.id])
+        self.assertEqual(
+            _task_run_scope_team_ids(payload),
+            [self.team.id],
+        )
 
         personal_org = Organization.objects.create(name="Personal Org")
         personal_team = Team.objects.create(organization=personal_org, name="Personal Team")
@@ -1631,12 +1632,15 @@ class TestExternalPRWebhook(TestCase):
         OrganizationMembership.objects.create(organization=personal_org, user=user)
         UserIntegration.objects.create(user=user, kind="github", integration_id="555000")
 
-        self.assertEqual(_task_run_scope_team_ids(payload), sorted([self.team.id, personal_team.id]))
+        self.assertEqual(
+            _task_run_scope_team_ids(payload),
+            sorted([self.team.id, personal_team.id]),
+        )
         # Attribution still resolves off the Integration rows alone.
         self.assertEqual(_installation_team_ids(payload), [self.team.id])
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_run_in_another_team_does_not_claim_the_delivery(self, mock_capture, mock_get_secret):
         # The run lookup is scoped to the installation's teams, so a run belonging to an
         # unrelated team cannot claim a PR URL it happens to share. Unscoped, the full-table
@@ -1674,8 +1678,8 @@ class TestExternalPRWebhook(TestCase):
         self.assertIsNone(properties["run_id"])
         self.assertEqual(_sample_value("posthog_tasks_github_webhook_task_run_lookup_total", labels), before + 1)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_external_pr_shared_installation_resolves_deterministically(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         other_team = Team.objects.create(organization=self.organization, name="Other External Team")
@@ -1690,8 +1694,8 @@ class TestExternalPRWebhook(TestCase):
         distinct_ids = {call[1]["distinct_id"] for call in mock_capture.call_args_list}
         self.assertEqual(distinct_ids, {str(expected_team.uuid)})
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.webhooks.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_external_pr_without_installation_block_is_dropped(self, mock_capture, mock_get_secret):
         mock_get_secret.return_value = self.webhook_secret
         payload = self._external_payload("opened", merged=False)
@@ -2181,7 +2185,7 @@ class TestGitHubWebhookFanout(TestCase):
         ]
     )
     @patch("products.conversations.backend.api.github_events.process_github_event")
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_issues_event_dispatched_to_conversations(self, _name, url, mock_secret, mock_task):
         mock_secret.return_value = self.webhook_secret
         mock_task.delay = MagicMock()
@@ -2204,7 +2208,7 @@ class TestGitHubWebhookFanout(TestCase):
         self.assertEqual(call_kwargs["repo"], "myorg/myrepo")
 
     @patch("products.conversations.backend.api.github_events.process_github_event")
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_issue_comment_event_dispatched_to_conversations(self, mock_secret, mock_task):
         mock_secret.return_value = self.webhook_secret
         mock_task.delay = MagicMock()
@@ -2225,7 +2229,7 @@ class TestGitHubWebhookFanout(TestCase):
         self.assertEqual(mock_task.delay.call_args[1]["event_type"], "issue_comment")
 
     @patch("products.conversations.backend.api.github_events.process_github_event")
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_issues_event_without_matching_team_returns_200(self, mock_secret, mock_task):
         mock_secret.return_value = self.webhook_secret
         mock_task.delay = MagicMock()
@@ -2249,8 +2253,8 @@ class TestGitHubWebhookFanout(TestCase):
             ("unified_url", "/webhooks/github/"),
         ]
     )
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
-    @patch("products.tasks.backend.models.posthoganalytics.capture")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.pull_requests.posthoganalytics.capture")
     def test_pull_request_routed(self, _name, url, mock_capture, mock_secret):
         mock_secret.return_value = self.webhook_secret
 
@@ -2268,7 +2272,7 @@ class TestGitHubWebhookFanout(TestCase):
         mock_capture.assert_called_once()
         self.assertEqual(mock_capture.call_args[1]["event"], "pr_merged")
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_unified_url_unknown_event_returns_200(self, mock_secret):
         mock_secret.return_value = self.webhook_secret
 
@@ -2277,7 +2281,7 @@ class TestGitHubWebhookFanout(TestCase):
 
         self.assertEqual(response.status_code, 200)
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_failed_handler_releases_dedup_so_redelivery_is_processed(self, mock_secret):
         # The dedup mark is set before the handler runs; a handler failure must release it so
         # GitHub's redelivery of the same GUID gets processed instead of silently skipped for
@@ -2304,7 +2308,7 @@ class TestGitHubWebhookFanout(TestCase):
             self.assertEqual(third.status_code, 200)
             mock_loops.assert_called_once()
 
-    @patch("products.tasks.backend.facade.webhooks.get_github_webhook_secret")
+    @patch("posthog.api.github_webhooks.views.get_github_webhook_secret")
     def test_unified_url_bad_signature_returns_403(self, mock_secret):
         mock_secret.return_value = self.webhook_secret
 
@@ -2486,7 +2490,7 @@ class TestAttributionDbAliases(TestCase):
     def test_default_only_when_no_replica_is_configured(self):
         self.assertEqual(_attribution_db_aliases(), ["default"])
 
-    @patch("products.tasks.backend.webhooks.router.db_for_read", return_value="default")
+    @patch("posthog.api.github_webhooks.attribution.router.db_for_read", return_value="default")
     def test_skips_a_configured_replica_the_router_would_not_read_from(self, _mock_db_for_read):
         # Bounding an alias means opening it, and connection setup is itself unbounded (these
         # aliases carry no connect_timeout), so a replica the router never reads from must not
@@ -2494,14 +2498,14 @@ class TestAttributionDbAliases(TestCase):
         with self._with_replica_configured():
             self.assertEqual(_attribution_db_aliases(), ["default"])
 
-    @patch("products.tasks.backend.webhooks.router.db_for_read", return_value="replica")
+    @patch("posthog.api.github_webhooks.attribution.router.db_for_read", return_value="replica")
     def test_skips_the_primary_when_every_model_reads_from_the_replica(self, _mock_db_for_read):
         # Symmetric to the above: a fully replica-opted deployment must not be made to wait on
         # the primary either, since opening it is just as unbounded.
         with self._with_replica_configured():
             self.assertEqual(_attribution_db_aliases(), ["replica"])
 
-    @patch("products.tasks.backend.webhooks.router.db_for_read")
+    @patch("posthog.api.github_webhooks.attribution.router.db_for_read")
     def test_covers_every_alias_the_models_read_from(self, mock_db_for_read):
         mock_db_for_read.side_effect = lambda model: "replica" if model is User else "default"
         with self._with_replica_configured():
