+++ products/tasks/backend/webhooks.py
@@ -1,39 +1,23 @@
-import hmac
-import uuid
-import hashlib
-from collections.abc import Iterator
-from contextlib import ExitStack, contextmanager
-
-from django.conf import settings
-from django.db import OperationalError, connections, router, transaction
-from django.db.backends.base.base import BaseDatabaseWrapper
+from django.db import transaction
 from django.db.models import Case, IntegerField, Q, Value, When
 from django.http import HttpResponse
 
 import structlog
-import posthoganalytics
-from social_django.models import UserSocialAuth
 
+from posthog.api.github_webhooks.contracts import PullRequestAttribution
+from posthog.api.github_webhooks.integrations import _SCOPE_DB_ALIAS, _installation_id, _installation_team_ids
+from posthog.api.github_webhooks.metrics import GitHubWebhookAnalyticsEvent, observe_github_webhook_pr_event_dropped
+from posthog.api.github_webhooks.pull_requests import capture_pr_event, pr_state_for_action
 from posthog.event_usage import groups
-from posthog.models.instance_setting import get_instance_setting
-from posthog.models.integration import Integration
 from posthog.models.organization import OrganizationMembership
 from posthog.models.team.team import Team
-from posthog.models.user import User
 from posthog.models.user_integration import UserIntegration
 
-from products.signals.backend.report_assignments import update_assignments_for_pull_request
-from products.signals.backend.report_generation.resolve_reviewers import resolve_org_github_login_to_users
+from products.signals.backend.facade.github import update_pull_request_assignments
 from products.tasks.backend.constants import PR_LOOP_ENABLED_STATE_KEY
 from products.tasks.backend.facade.api import post_pr_created_thread_update, signal_workflow_completion
 from products.tasks.backend.facade.cancellation import cancel_task_run
-from products.tasks.backend.metrics import (
-    GitHubWebhookAnalyticsEvent,
-    GitHubWebhookAttributionOutcome,
-    observe_github_webhook_attribution,
-    observe_github_webhook_pr_event_dropped,
-    observe_github_webhook_task_run_lookup,
-)
+from products.tasks.backend.metrics import observe_github_webhook_task_run_lookup
 from products.tasks.backend.models import Task, TaskRun
 from products.tasks.backend.pr_urls import merge_pr_output, read_pr_urls
 from products.tasks.backend.prompts import WIZARD_HEAD_BRANCH_PREFIX
@@ -186,53 +170,34 @@ def find_task_run(
     return None
 
 
-def verify_github_signature(payload: bytes, signature: str | None, secret: str) -> bool:
-    """
-    Verify the GitHub webhook signature using HMAC-SHA256.
-
-    GitHub sends a signature in the X-Hub-Signature-256 header in the format:
-    sha256=<hex_digest>
-    """
-    if not signature or not signature.startswith("sha256="):
-        return False
-
-    expected_signature = (
-        "sha256="
-        + hmac.new(
-            secret.encode("utf-8"),
-            payload,
-            hashlib.sha256,
-        ).hexdigest()
-    )
-
-    return hmac.compare_digest(expected_signature, signature)
-
-
-def get_github_webhook_secret() -> str | None:
-    """Get the GitHub webhook secret from instance settings."""
-    secret = get_instance_setting("GITHUB_WEBHOOK_SECRET")
-    return secret if secret else None
-
-
-def _pr_state_for_action(action: str | None, pull_request: dict) -> str | None:
-    """The ``output.pr_state`` a webhook action moves a run's PR to, in the
-    same open/draft/merged/closed vocabulary the GitHub snapshot uses. None
-    for actions that don't change the state (comments, labels, pushes)."""
-    if action in ("opened", "reopened"):
-        return "draft" if pull_request.get("draft") else "open"
-    if action == "ready_for_review":
-        return "open"
-    if action == "converted_to_draft":
-        return "draft"
-    if action == "closed":
-        return "merged" if pull_request.get("merged") else "closed"
-    return None
+def _capture_task_pr_event(payload: dict, task_run: TaskRun | None, event: GitHubWebhookAnalyticsEvent) -> None:
+    try:
+        attribution = None
+        if task_run is not None:
+            attribution = PullRequestAttribution(
+                source="task",
+                team_id=task_run.team_id,
+                distinct_id=(
+                    str(task_run.task.created_by.distinct_id)
+                    if task_run.task.created_by_id and task_run.task.created_by
+                    else str(task_run.team.uuid)
+                ),
+                groups=groups(team=task_run.team),
+                properties=task_run.analytics_properties(),
+                include_content=True,
+                send_feature_flags=True,
+            )
+    except Exception as error:
+        observe_github_webhook_pr_event_dropped(analytics_event=event, reason="capture_exception")
+        logger.warning("github_pr_webhook_capture_failed", analytics_event=event, error=str(error))
+        return
+    capture_pr_event(payload, attribution, event)
 
 
 def handle_pull_request_event(payload: dict) -> HttpResponse:
     """Process a pre-verified pull_request webhook event.
 
-    Called from ``posthog.urls.github_webhook`` (unified dispatcher).
+    Called from the shared GitHub webhook dispatcher (unified dispatcher).
     """
     action = payload.get("action")
     pull_request = payload.get("pull_request", {})
@@ -243,7 +208,7 @@ def handle_pull_request_event(payload: dict) -> HttpResponse:
         logger.warning("github_pr_webhook_no_pr_url", action=action)
         return HttpResponse(status=200)
 
-    pr_state = _pr_state_for_action(action, pull_request)
+    pr_state = pr_state_for_action(action, pull_request)
     analytics_event: GitHubWebhookAnalyticsEvent | None = None
     if action == "opened":
         event_action = "created"
@@ -267,7 +232,6 @@ def handle_pull_request_event(payload: dict) -> HttpResponse:
     branch = pull_request.get("head", {}).get("ref")
     repository_full_name = (payload.get("repository") or {}).get("full_name")
     scoped_team_ids = _task_run_scope_team_ids(payload)
-    assignment_team_ids = _installation_team_ids(payload)
     task_run = find_task_run(pr_url=pr_url, branch=branch, repository=repository_full_name, team_ids=scoped_team_ids)
     claimed_pr_urls = (
         read_pr_urls(task_run.output if isinstance(task_run.output, dict) else {}) if task_run is not None else []
@@ -311,23 +275,10 @@ def handle_pull_request_event(payload: dict) -> HttpResponse:
     ):
         _record_run_pr_state(task_run, pr_state)
 
-    if pr_state is not None and repository_full_name and assignment_team_ids:
-        try:
-            update_assignments_for_pull_request(
-                team_ids=assignment_team_ids,
-                repository=repository_full_name,
-                pr_number=int(pull_request.get("number")),
-                pr_state=pr_state,
-            )
-        except (TypeError, ValueError):
-            logger.warning("github_pr_webhook_signal_assignment_missing_number", pr_url=pr_url)
-        except Exception:
-            logger.exception("github_pr_webhook_signal_assignment_update_failed", pr_url=pr_url)
+    update_pull_request_assignments(payload, pr_state)
 
     if analytics_event is not None:
-        # Deterministic UUID dedupes duplicate webhook deliveries of the same PR action.
-        event_uuid = str(uuid.uuid5(uuid.NAMESPACE_URL, f"{pr_url}:{analytics_event}"))
-        _capture_pr_event(payload, task_run, analytics_event, event_uuid)
+        _capture_task_pr_event(payload, task_run, analytics_event)
 
     if action == "closed" and merged:
         # Only trust the merge for the run that actually claims this PR URL. The pr_url backstop
@@ -347,7 +298,7 @@ def handle_pull_request_event(payload: dict) -> HttpResponse:
 def handle_pull_request_review_event(payload: dict) -> HttpResponse:
     """Process a pre-verified pull_request_review webhook event.
 
-    Called from ``posthog.urls.github_webhook`` (unified dispatcher). Captures a
+    Called from the shared GitHub webhook dispatcher (unified dispatcher). Captures a
     ``pr_reviewed`` analytics event for human review submissions (approved,
     changes_requested, commented), attributed to the reviewer when their GitHub
     login resolves to an org member.
@@ -375,9 +326,7 @@ def handle_pull_request_review_event(payload: dict) -> HttpResponse:
         pr_url=pr_url, branch=branch, repository=repository_full_name, team_ids=_task_run_scope_team_ids(payload)
     )
 
-    # One review submission = one event; GitHub redeliveries collapse on the review id.
-    event_uuid = str(uuid.uuid5(uuid.NAMESPACE_URL, f"{pr_url}:pr_reviewed:{review.get('id')}"))
-    _capture_pr_review_event(payload, task_run, event_uuid)
+    _capture_task_pr_event(payload, task_run, "pr_reviewed")
 
     logger.info(
         "github_pr_review_webhook_processed",
@@ -568,369 +517,6 @@ def _record_run_output_field(task_run: TaskRun, key: str, value: str | bool, fai
         return False
 
 
-# Nulled on external PRs so their schema matches task-originated PR events.
-_TASK_ATTRIBUTION_KEYS = ("task_id", "run_id", "origin_product", "signal_report_id", "environment", "mode", "title")
-
-# What a PR says, rather than how big it is. Only task-authored PRs carry these values -- an
-# external PR's own words are customer business context, so it gets the keys as nulls for
-# schema parity, the same way _TASK_ATTRIBUTION_KEYS works.
-_PR_CONTENT_KEYS = (
-    "pr_title",
-    "pr_body",
-    "pr_body_truncated",
-    "pr_labels",
-    "pr_requested_reviewers",
-    "pr_is_draft",
-)
-
-# A PR body is the biggest string on the delivery, and an agent-authored one can run long.
-# Cap it so one verbose body cannot push the event past the capture size limit.
-_PR_BODY_MAX_CHARS = 10_000
-
-
-def _account_type(payload: dict) -> str | None:
-    """Whether the webhook's repo is owned by a GitHub org or a personal account.
-
-    ``repository.owner.type`` is "Organization" or "User"; the top-level
-    ``organization`` object is present only for org-owned repos and backs it up
-    when the owner block is missing. Returns None when neither signal is present.
-    """
-    owner_type = ((payload.get("repository") or {}).get("owner") or {}).get("type")
-    if owner_type == "Organization":
-        return "organization"
-    if owner_type == "User":
-        return "personal"
-    if payload.get("organization"):
-        return "organization"
-    return None
-
-
-def _pr_payload_properties(payload: dict) -> dict:
-    pull_request = payload.get("pull_request") or {}
-    return {
-        "pr_url": pull_request.get("html_url"),
-        "pr_number": pull_request.get("number"),
-        "pr_author": (pull_request.get("user") or {}).get("login"),
-        "pr_base_ref": (pull_request.get("base") or {}).get("ref"),
-        "pr_head_ref": (pull_request.get("head") or {}).get("ref"),
-        "pr_additions": pull_request.get("additions"),
-        "pr_deletions": pull_request.get("deletions"),
-        "pr_changed_files": pull_request.get("changed_files"),
-        "pr_commits": pull_request.get("commits"),
-        "account_type": _account_type(payload),
-        "repo_owner_type": ((payload.get("repository") or {}).get("owner") or {}).get("type"),
-    }
-
-
-def _pr_content_properties(payload: dict) -> dict:
-    """What a task-authored PR says: its title, body, labels, requested reviewers, draft state."""
-    pull_request = payload.get("pull_request") or {}
-    body = pull_request.get("body") or ""
-    return {
-        "pr_title": pull_request.get("title"),
-        "pr_body": body[:_PR_BODY_MAX_CHARS],
-        "pr_body_truncated": len(body) > _PR_BODY_MAX_CHARS,
-        "pr_labels": [label.get("name") for label in (pull_request.get("labels") or []) if label.get("name")],
-        "pr_requested_reviewers": [
-            reviewer.get("login")
-            for reviewer in (pull_request.get("requested_reviewers") or [])
-            if reviewer.get("login")
-        ],
-        "pr_is_draft": pull_request.get("draft"),
-    }
-
-
-# Cap the org-member lookup that attributes the merger (and reviewer). GitHub gives a
-# pull_request delivery one short window and never retries it, and the merged branch runs
-# functional side effects (merge bookkeeping, signal-report resolution, wizard wind-down)
-# right after capture. A slow lookup on the request path can therefore cost the whole
-# delivery, not just the analytics event. Bounding it degrades to no attribution instead.
-_ATTRIBUTION_STATEMENT_TIMEOUT_MS = 800
-
-# Models the org-member resolver reads. ReplicaRouter only sends a model to the replica when
-# that model is named in READ_REPLICA_OPT_IN, so the set of aliases the lookup can touch is
-# knowable up front.
-_ATTRIBUTION_MODELS = (Team, User, OrganizationMembership, UserSocialAuth, UserIntegration, Integration)
-
-
-def _attribution_db_aliases() -> list[str]:
-    """The aliases the org-member lookup actually reads from, deduped, in model order.
-
-    Bounding an alias means opening it, and opening is itself unbounded -- ``postgres_config``
-    sets no ``connect_timeout`` on these aliases. So take the set from the router rather than
-    assuming: reaching for an alias the resolver never uses could stall the webhook on
-    connection setup before the cap is installed, which is the failure this exists to prevent.
-    That cuts both ways -- a fully replica-opted deployment must not be made to wait on the
-    primary either.
-    """
-    aliases: list[str] = []
-    for model in _ATTRIBUTION_MODELS:
-        alias = router.db_for_read(model) or "default"
-        if alias not in aliases and alias in settings.DATABASES:
-            aliases.append(alias)
-    return aliases
-
-
-def _read_statement_timeout(connection: BaseDatabaseWrapper) -> str | None:
-    with connection.cursor() as cursor:
-        cursor.execute("SHOW statement_timeout")
-        row = cursor.fetchone()
-    return row[0] if row else None
-
-
-def _apply_statement_timeout(connection: BaseDatabaseWrapper, value: str) -> None:
-    # set_config(..., is_local=True) is SET LOCAL, but takes the value as a bind parameter,
-    # so a restored value ("30s", "0", ...) does not have to be quoted by hand.
-    with connection.cursor() as cursor:
-        cursor.execute("SELECT set_config('statement_timeout', %s, true)", [value])
-
-
-@contextmanager
-def _statement_timeout(connection: BaseDatabaseWrapper, timeout_ms: int, *, restore: bool) -> Iterator[None]:
-    """Cap statements on one connection, optionally putting the previous value back."""
-    previous = _read_statement_timeout(connection) if restore else None
-    _apply_statement_timeout(connection, f"{timeout_ms}ms")
-
-    yield
-
-    # Only reached when the block succeeded. If it raised, the enclosing atomic() rolls the
-    # (sub)transaction back and PostgreSQL undoes SET LOCAL with it, so there is nothing to
-    # restore -- and a statement on an aborted transaction would error anyway.
-    if previous:
-        _apply_statement_timeout(connection, previous)
-
-
-@contextmanager
-def _bounded_attribution_lookup() -> Iterator[None]:
-    """Run a block under a per-statement timeout on each configured DB the lookup may use.
-
-    A read routed to an alias joins that alias's open transaction, so ``SET LOCAL
-    statement_timeout`` there caps the query regardless of read-replica routing.
-    """
-    with ExitStack() as stack:
-        for alias in _attribution_db_aliases():
-            connection = connections[alias]
-            # SET LOCAL dies with the transaction it was set in, so the cap only needs
-            # restoring when we are joining a transaction somebody else owns -- a future
-            # caller wrapping this in its own atomic block, or ATOMIC_REQUESTS (which
-            # PostHog does not enable today). Otherwise the commit below ends it for us.
-            restore = connection.in_atomic_block
-            stack.enter_context(transaction.atomic(using=alias))
-            stack.enter_context(_statement_timeout(connection, _ATTRIBUTION_STATEMENT_TIMEOUT_MS, restore=restore))
-        yield
-
-
-# PostgreSQL raises query_canceled when statement_timeout fires. Django wraps the driver
-# error in OperationalError, so the SQLSTATE lives on the cause -- psycopg3 spells it
-# `sqlstate`, psycopg2 `pgcode`. The message check is the fallback for anything that loses
-# the cause on the way up.
-_QUERY_CANCELED_SQLSTATE = "57014"
-
-
-def _is_statement_timeout(error: Exception) -> bool:
-    if not isinstance(error, OperationalError):
-        return False
-    cause = error.__cause__
-    if getattr(cause, "sqlstate", None) == _QUERY_CANCELED_SQLSTATE:
-        return True
-    if getattr(cause, "pgcode", None) == _QUERY_CANCELED_SQLSTATE:
-        return True
-    return "statement timeout" in str(error).lower()
-
-
-def _resolve_github_login_distinct_id(login: str | None, team_id: int) -> str | None:
-    """Distinct id of the org member matching a GitHub login, or None when unresolvable.
-
-    Runs under a per-statement timeout so a slow member lookup cannot hold the webhook
-    open past GitHub's delivery timeout (see ``_ATTRIBUTION_STATEMENT_TIMEOUT_MS``).
-    """
-    if not login:
-        return None
-    try:
-        with _bounded_attribution_lookup():
-            resolved = resolve_org_github_login_to_users(team_id, [login]).get(str(login).strip().lower())
-    except Exception as e:
-        # timeout is meant to be the leading indicator for the cap we just installed, so it
-        # has to mean "statement cancelled", not "any OperationalError" -- connection resets
-        # and other DB incidents raise the same class and would drown the signal.
-        outcome: GitHubWebhookAttributionOutcome = "timeout" if _is_statement_timeout(e) else "error"
-        observe_github_webhook_attribution(outcome=outcome)
-        logger.warning(
-            "github_webhook_login_resolution_failed", login=login, team_id=team_id, outcome=outcome, error=str(e)
-        )
-        return None
-    if resolved is None:
-        observe_github_webhook_attribution(outcome="unresolved")
-        return None
-    observe_github_webhook_attribution(outcome="resolved")
-    return str(resolved.distinct_id)
-
-
-def _merged_by_attribution(payload: dict, team_id: int) -> tuple[dict, str | None]:
-    """Identity of the GitHub user who merged the PR, resolved to a PostHog user when possible.
-
-    Merging is the one unambiguous personal act in the loop, so when the merger's GitHub
-    login maps to an org member the pr_merged event attributes to them. Without a match the
-    event keeps the task's assigned user (an auto-resolved reviewer or fallback for
-    auto-started reports), so a consumer tells the two apart by the presence of
-    pr_merged_by_distinct_id.
-    """
-    merged_by = (payload.get("pull_request") or {}).get("merged_by") or {}
-    login = merged_by.get("login")
-    if not login:
-        return {}, None
-    properties: dict = {"pr_merged_by_login": login, "pr_merged_by_id": merged_by.get("id")}
-    distinct_id = _resolve_github_login_distinct_id(login, team_id)
-    if distinct_id is not None:
-        properties["pr_merged_by_distinct_id"] = distinct_id
-    return properties, distinct_id
-
-
-def _capture_pr_review_event(payload: dict, task_run: TaskRun | None, event_uuid: str) -> None:
-    review = payload.get("review") or {}
-    reviewer = review.get("user") or {}
-    login = reviewer.get("login")
-    review_properties: dict = {
-        "pr_review_state": review.get("state"),
-        "pr_reviewed_by_login": login,
-        "pr_reviewed_by_id": reviewer.get("id"),
-    }
-    pr_properties = {**_pr_payload_properties(payload), **review_properties}
-
-    if task_run is not None:
-        reviewer_distinct_id = _resolve_github_login_distinct_id(login, task_run.team_id)
-        if reviewer_distinct_id is not None:
-            pr_properties["pr_reviewed_by_distinct_id"] = reviewer_distinct_id
-        captured = task_run.capture_event(
-            "pr_reviewed",
-            {**pr_properties, **_pr_content_properties(payload), "pr_source": "task"},
-            event_uuid=event_uuid,
-            distinct_id_override=reviewer_distinct_id,
-        )
-        if not captured:
-            observe_github_webhook_pr_event_dropped(analytics_event="pr_reviewed", reason="capture_exception")
-        return
-
-    team = _resolve_external_team(payload)
-    if team is None:
-        observe_github_webhook_pr_event_dropped(analytics_event="pr_reviewed", reason="unresolved_installation")
-        logger.debug("github_pr_review_webhook_unresolved_installation", pr_url=pr_properties.get("pr_url"))
-        return
-
-    reviewer_distinct_id = _resolve_github_login_distinct_id(login, team.id)
-    if reviewer_distinct_id is not None:
-        pr_properties["pr_reviewed_by_distinct_id"] = reviewer_distinct_id
-
-    properties: dict = {
-        **pr_properties,
-        "repository": ((payload.get("repository") or {}).get("full_name") or "").strip().lower() or None,
-        "pr_source": "external",
-        "team_id": team.id,
-        # title and PR content omitted to avoid leaking customer business context.
-        **dict.fromkeys(_TASK_ATTRIBUTION_KEYS, None),
-        **dict.fromkeys(_PR_CONTENT_KEYS, None),
-    }
-
-    try:
-        posthoganalytics.capture(
-            distinct_id=reviewer_distinct_id or str(team.uuid),
-            event="pr_reviewed",
-            properties=properties,
-            groups=groups(team=team),
-            uuid=event_uuid,
-        )
-    except Exception as e:
-        observe_github_webhook_pr_event_dropped(analytics_event="pr_reviewed", reason="capture_exception")
-        logger.warning("github_pr_review_webhook_capture_failed", error=str(e))
-
-
-def _capture_pr_event(
-    payload: dict, task_run: TaskRun | None, analytics_event: GitHubWebhookAnalyticsEvent, event_uuid: str
-) -> None:
-    pr_properties = _pr_payload_properties(payload)
-
-    if task_run is not None:
-        merger_distinct_id: str | None = None
-        if analytics_event == "pr_merged":
-            merged_by_properties, merger_distinct_id = _merged_by_attribution(payload, task_run.team_id)
-            pr_properties = {**pr_properties, **merged_by_properties}
-        captured = task_run.capture_event(
-            analytics_event,
-            {**pr_properties, **_pr_content_properties(payload), "pr_source": "task"},
-            event_uuid=event_uuid,
-            distinct_id_override=merger_distinct_id,
-        )
-        if not captured:
-            observe_github_webhook_pr_event_dropped(analytics_event=analytics_event, reason="capture_exception")
-        return
-
-    team = _resolve_external_team(payload)
-    if team is None:
-        observe_github_webhook_pr_event_dropped(analytics_event=analytics_event, reason="unresolved_installation")
-        logger.debug("github_pr_webhook_unresolved_installation", pr_url=pr_properties.get("pr_url"))
-        return
-
-    external_distinct_id = str(team.uuid)
-    if analytics_event == "pr_merged":
-        merged_by_properties, merger_distinct_id = _merged_by_attribution(payload, team.id)
-        pr_properties = {**pr_properties, **merged_by_properties}
-        external_distinct_id = merger_distinct_id or external_distinct_id
-
-    properties: dict = {
-        **pr_properties,
-        "repository": ((payload.get("repository") or {}).get("full_name") or "").strip().lower() or None,
-        "pr_source": "external",
-        "team_id": team.id,
-        # title and PR content omitted to avoid leaking customer business context.
-        **dict.fromkeys(_TASK_ATTRIBUTION_KEYS, None),
-        **dict.fromkeys(_PR_CONTENT_KEYS, None),
-    }
-
-    try:
-        posthoganalytics.capture(
-            distinct_id=external_distinct_id,
-            event=analytics_event,
-            properties=properties,
-            groups=groups(team=team),
-            uuid=event_uuid,
-        )
-    except Exception as e:
-        observe_github_webhook_pr_event_dropped(analytics_event=analytics_event, reason="capture_exception")
-        logger.warning("github_pr_webhook_capture_failed", analytics_event=analytics_event, error=str(e))
-
-
-def _installation_id(payload: dict) -> str | None:
-    """The delivery's GitHub App installation id, in the form the integration rows store it."""
-    installation_id = (payload.get("installation") or {}).get("id")
-    return None if installation_id is None else str(installation_id)
-
-
-# The run lookup these feed reads TaskRun off the writer, and they run on the request path
-# outside the bounded attribution block. Pin them to the writer too: a replica-opted
-# Integration or Team would otherwise let a slow replica stall a delivery whose own lookup
-# never needed it, and replica lag could hide a freshly connected installation.
-_SCOPE_DB_ALIAS = "default"
-
-
-def _installation_team_ids(payload: dict) -> list[int]:
-    """Teams whose GitHub Integration matches the delivery's installation, in deterministic order.
-
-    Empty when the payload carries no installation id or no Integration matches it — the
-    lookups that take this fall back to their unscoped behaviour in that case.
-    """
-    external_id = _installation_id(payload)
-    if external_id is None:
-        return []
-
-    # One installation can map to multiple teams; order_by makes attribution deterministic.
-    return list(
-        Integration.objects.using(_SCOPE_DB_ALIAS)
-        .filter(kind="github", integration_id=external_id)
-        .order_by("team_id")
-        .values_list("team_id", flat=True)
-    )
-
-
 def _task_run_scope_team_ids(payload: dict) -> list[int]:
     """Teams to scope the TaskRun lookup to, or empty to leave the lookup unscoped.
 
@@ -968,10 +554,3 @@ def _task_run_scope_team_ids(payload: dict) -> list[int]:
     )
 
     return sorted(team_ids)
-
-
-def _resolve_external_team(payload: dict) -> Team | None:
-    team_ids = _installation_team_ids(payload)
-    if not team_ids:
-        return None
-    return Team.objects.filter(pk=team_ids[0]).first()
